fix: the guard that was supposed to make an unnamed error code impossible read one file out of nine
An adversarial review of tonight's five merged PRs, run through omp on three
scopes and verified by hand. Six findings held up; the two worst are fixed
here, the rest are corrections and three issues.

**The guard was not a guard.** worker/tests/outcomes.test.ts was introduced in
#84 to make it impossible for the relay to answer with a code no app has a
sentence for. It scanned the text of src/app.ts, one file, while
invalid-coordinate and invalid-description are thrown from src/domain.ts on
the way through createReport and reach a phone exactly like the rest. Both
were missing from data/relay-outcomes.json and the build was green. It reads
every source file now, and it also refuses a code it cannot READ: a code
built from a variable would have passed every check by being invisible, so
that fails instead. The two codes have sentences.

**Pressing send again after a failed attempt queued a second copy of the
same report.** The button is disabled only while a delivery runs, so the
press is available the moment an attempt gives up, and sendToRelay enqueued
unconditionally. Both copies then went out when the network returned. A
press on a report that is already queued is now a retry of that report. This
is not #85, which is the same duplicate after a SUCCESS and is still open.

**The message promised the report was waiting when it was not.** On the
fallback path, where enqueue itself failed, the app still said the ábending
was waiting in the phone and would go when the network came back. Nothing
had been written down and nothing retried it, so the assurance was exactly
what would make somebody leave the screen and lose the report. That path
says so now.

Android: a permission question that cannot be asked now fails towards the
retry that might work rather than towards telling somebody their own device
is locked; a non-string error field reads as no code, matching Swift; and
atMs clamps before narrowing, so a session past 24.8 days cannot wrap
negative and clamp to zero.

Three comments described the wrong failure mode and are corrected rather
than deleted, including one on both platforms that had it exactly backwards:
a transport that never reports does not leave the channel idle, it wedges it
for the session. BatchReceipt now enforces exactly-once instead of asking
for it.

The record was wrong in three places and is corrected in the file and in
#85: the session has three photo-captured events rather than one, the five
screenshots came from both testers rather than one, and the session-start
estimate rests on one batch whose sibling disagrees by seven seconds.

Deferred with reasons: #88 no idempotency key, so a duplicate cannot be
caught downstream; #89 both app-layer state machines untested, so #76 would
revert green; #90 the field-test checker validates only the timeline.

+++ android/app/src/main/kotlin/is/borgarland/poc/net/Telemetry.kt
@@ -126,8 +126,14 @@ class Telemetry {
      * Injectable so a unit test can capture the body without a network.
      *
      * The closure must invoke its completion exactly once. Invoking it late is
-     * fine and is the normal case; never invoking it leaves the channel with no
-     * flush in flight and a buffer that only drains at the cap.
+     * fine and is the normal case.
+     *
+     * Never invoking it wedges the channel for the rest of the session:
+     * [finish] is the only place `inFlight` is cleared, so every later flush
+     * returns at once and the buffer never drains at all. [track] keeps shedding
+     * its oldest event at the cap, so events are lost silently, which is the
+     * hole #74 was about wearing a different hat. Failures here are swallowed by
+     * design, so nothing would ever say so.
      */
     var send: ((String, (BatchOutcome) -> Unit) -> Unit)? = null
 
@@ -156,7 +162,7 @@ class Telemetry {
     /** Records an event. Never throws and never blocks on the network: the
      * buffer is capped, and the flush returns before its request does. */
     fun track(event: TelemetryEvent) {
-        val atMs = clampAtMs((System.currentTimeMillis() - sessionStartMs).toInt())
+        val atMs = clampAtMs(System.currentTimeMillis() - sessionStartMs)
         var shouldFlush = false
         synchronized(lock) {
             // description-length is emitted on every keystroke; coalesce
@@ -277,9 +283,16 @@ class Telemetry {
         }
     }
 
-    /** A day in milliseconds; an offset beyond that is a broken clock, not a
-     * session (the Worker's own bound). */
-    private fun clampAtMs(ms: Int): Int = ms.coerceIn(0, MAX_AT_MS)
+    /**
+     * A day in milliseconds; an offset beyond that is a broken clock, not a
+     * session (the Worker's own bound).
+     *
+     * Takes a Long and clamps before narrowing. Truncating first would wrap a
+     * session past about 24.8 days to a negative Int and clamp it to 0, which
+     * defeats the bound instead of enforcing it. Unreachable for a camera app,
+     * and free to get right.
+     */
+    private fun clampAtMs(ms: Long): Int = ms.coerceIn(0L, MAX_AT_MS.toLong()).toInt()
 }
 
 /**
